Fold McpGatewayRoutes into McpJsonRpcHandler
The route mounting had one consumer and no logic of its own, so the
handler now implements SuppliesGatewayRoutes directly: it mounts
POST /mcp, authenticates the request, and dispatches it in one class.
